backend/requirements.txt
django>=4.2,<5.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
django-ratelimit
scikit-learn
pandas
joblib
gunicorn
python-dotenv
whitenoise